pyproject.toml
[project]
name = "random-image"
version = "1.0.0"
description = "随机图片服务 - 提供随机图片获取和智能压缩功能的FastAPI服务"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.128.2",
    "pillow>=12.1.0",
    "uvicorn>=0.40.0",
    "python-multipart>=0.0.17",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "ruff>=0.8.0",
]

test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
include = ["random_image*"]

[tool.setuptools.package-dir]
"" = "src"

[tool.uv]
package = true

[tool.black]
line-length = 88
target-version = ['py311']

[tool.ruff]
line-length = 88
target-version = "py311"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # mccabe
    "B",  # flake8-bugbear
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src/random_image"]
omit = ["*/tests/*"]

[project.scripts]
random-image = "random_image.main:app"
